% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\name{dimnames,initRandrot-method}
\alias{dimnames,initRandrot-method}
\alias{dimnames,initBatchRandrot-method}
\title{Dimnames of an Object}
\usage{
\S4method{dimnames}{initRandrot}(x)

\S4method{dimnames}{initBatchRandrot}(x)
}
\arguments{
\item{x}{An object of class
\code{\link[randRotation:initRandrot-class]{initRandrot-class}} or
\code{\link[randRotation:initBatchRandrot-class]{initBatchRandrot-class}}.}
}
\value{
A \code{list} with names of \code{features} and \code{samples}, see
  \code{\link[randRotation:initRandrot]{initRandrot}}.
}
\description{
Retrieve the dimnames of an object.
}
