% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\docType{class}
\name{init.batch.randrot-class}
\alias{init.batch.randrot-class}
\title{Initialised random rotation batch object}
\description{
This class contains \code{init.randrot} or \code{init.randrot.w} class
objects for each batch. See also descriptions in
\code{\link[randRotation:init.randrot]{init.randrot}} and
\code{\link[randRotation:init.randrot-class]{init.randrot-class}}.
}
\section{Components}{

\describe{
  \item{\code{batch.obj}}{List of \code{init.randrot} or \code{init.randrot.w} class objects for each batch.}
  \item{\code{split.by}}{List of sample indices for each batch.}
}
}

\author{
Peter Hettegger
}
