% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\docType{class}
\name{initBatchRandrot-class}
\alias{initBatchRandrot-class}
\title{Initialised random rotation batch object}
\description{
This class contains \code{initRandrot} or \code{initRandrotW} class
objects for each batch. See also descriptions in
\code{\link[randRotation:initRandrot]{initRandrot}} and
\code{\link[randRotation:initRandrot-class]{initRandrot-class}}.
}
\section{Components}{

\describe{
  \item{\code{batch.obj}}{List of \code{initRandrot} or \code{initRandrotW} class objects for each batch.}
  \item{\code{split.by}}{List of sample indices for each batch.}
}
}

\author{
Peter Hettegger
}
