% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\name{randorth}
\alias{randorth}
\title{Random rotation matrix}
\usage{
randorth(n, type = c("orthonormal", "unitary"), I.matrix = FALSE)
}
\arguments{
\item{n}{\code{numeric} of length 1 defining the dimensions of the \code{n x n} square matrix.}

\item{type}{Either \code{"orthonormal"} or \code{"unitary"} defining whether a real orthonormal matrix or a complex unitary matrix should be returned.}

\item{I.matrix}{If \code{TRUE}, the identity matrix is returned.}
}
\value{
A random rotation matrix of dimension \code{n x n}.
}
\description{
Generation of a random \code{n x n} rotation matrix (random orthogonal matrix) distributed with haar measure.
}
\details{
Adapted from pracma package (\code{\link[pracma:randortho]{pracma::randortho}})
\insertCite{Stewart1980a}{randRotation}
}
\examples{

# The following example shows the orthogonality of the random rotation matrix:
R1 = randorth(4)
zapsmall(t(R1) \%*\% R1)

R1 = randorth(4, "unitary")
zapsmall(Conj(t(R1)) \%*\% R1)

# The following example shows the distribution of 2-dimensional random orthogonal vectors
# on the unit circle.
tmp1 = vapply(1:200, function(i)randorth(2)[,1], numeric(2))
plot(t(tmp1), xlab = "x", ylab = "y")

}
\references{
\insertAllCited{}
}
\author{
Peter Hettegger
}
