% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\name{randpermut}
\alias{randpermut}
\title{Generate random permutation matrix for n samples}
\usage{
randpermut(n)
}
\arguments{
\item{n}{Number of samples}
}
\value{
A random permutation matrix of dimension \code{n x n}
}
\description{
Generate a random permutation matrix for \code{n} samples.
}
\details{
This methods generates an orthogonal matrix with only
  one entry in each row and column being \code{1}, all other entries being
  \code{0}.
}
\examples{
tmp1 <- randpermut(5)
t(tmp1) \%*\% tmp1
}
\author{
Peter Hettegger
}
