% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\docType{methods}
\name{randrot}
\alias{randrot}
\alias{randrot,init.randrot-method}
\alias{randrot,init.randrot.w-method}
\alias{randrot,init.batch.randrot-method}
\title{Random rotation of initialised object}
\usage{
randrot(object, ...)

\S4method{randrot}{init.randrot}(object, ...)

\S4method{randrot}{init.randrot.w}(object, ...)

\S4method{randrot}{init.batch.randrot}(object, ...)
}
\arguments{
\item{object}{An initialised object of class \code{\link[randRotation:init.randrot-class]{init.randrot-class}} or \code{\link[randRotation:init.batch.randrot-class]{init.batch.randrot-class}}.}

\item{...}{further arguments passed to \code{\link[randRotation:randorth]{randorth}}}
}
\value{
\code{numeric} matrix of rotated data under the specified combined null hypothesis.
}
\description{
Perform random data rotation of a previously initialised object (see \code{\link[randRotation:init.randrot]{init.randrot}}) associated with the null hypothesis \eqn{H_{0}: \beta_{coef.h} = 0}{H0: \beta_coef.h = 0}.
}
\details{
This function generates a randomly rotated dataset from an initialised randrot object (see \code{\link[randRotation:init.randrot]{init.randrot}}). See also package vignette for application examples.
Only the numerical matrix of rotated data is returned, no design matrix, weights or other info is return for efficiency purposes. Please consider that
if you e.g. use \code{weights} or if you use \code{\link[randRotation:rotate.stat]{rotate.stat}}, you may need to forward the design matrix \code{X}, \code{weights} etc. to subsequent anylyses.
See the example in \code{\link[randRotation:rotate.stat]{rotate.stat}}.

Details on the calculation of a rotated datset are given in \code{\link[randRotation:init.randrot]{init.randrot}} and \insertCite{Langsrud2005}{randRotation}.
}
\examples{
# See '?rotate.stat'
}
\references{
\insertAllCited{}
}
\author{
Peter Hettegger
}
