\name{randRotation-deprecated}
\alias{randRotation-deprecated}
\title{Deprecated functions in package \sQuote{randRotation}}

\description{
    These functions are provided for compatibility with older versions
    of \sQuote{randRotation} only, and will be defunct at the next release.
}

\details{

    The following functions are deprecated and will be made defunct; use
    the replacement indicated below:
        \itemize{

            \item{df_estimate: There is no replacement function for df_estimate.}

        }
}
