% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\docType{methods}
\name{show,init.randrot-method}
\alias{show,init.randrot-method}
\alias{show,init.batch.randrot-method}
\title{Show an Object}
\usage{
\S4method{show}{init.randrot}(object)

\S4method{show}{init.batch.randrot}(object)
}
\arguments{
\item{object}{An object of class \code{\link[randRotation:init.randrot-class]{init.randrot-class}} or \code{\link[randRotation:init.batch.randrot-class]{init.batch.randrot-class}}.}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
Display the object by printing structured summary information.
}
\details{
The show method always displays the original design matrix (\code{X}), not the transformed (whitened) versions.
}
