% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\name{weights,initRandrot-method}
\alias{weights,initRandrot-method}
\alias{weights,initBatchRandrot-method}
\title{Extract model weights}
\usage{
\S4method{weights}{initRandrot}(object, ...)

\S4method{weights}{initBatchRandrot}(object, ...)
}
\arguments{
\item{object}{An object of class
\code{\link[randRotation:initRandrot-class]{initRandrot-class}},
\code{\link[randRotation:initRandrot-class]{initRandrotW-class}} or
\code{\link[randRotation:initBatchRandrot-class]{initBatchRandrot-class}}.}

\item{...}{Kept for compatibility with the default method, see
\code{?stats::\link[stats:weights]{weights}}.
For objects defined in package \link[randRotation:randRotation-package]{randRotation},
this argument is currently not needed.}
}
\value{
Weights extracted from the object \code{object}. \code{NULL} if no
weights were specified.
See \code{?stats::\link[stats:weights]{weights}} for the value returned by the default method.
}
\description{
\code{weights} is a generic function which extracts fitting weights from objects returned by modeling functions.
NOTE: This man page is for the weights S4 generic function defined in the \link[randRotation:randRotation-package]{randRotation} package.
}
\examples{
weights
showMethods("weights")
selectMethod("weights", "ANY")  # the default method

}
