% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\docType{methods}
\name{dim,init.randrot-method}
\alias{dim,init.randrot-method}
\alias{dim,init.batch.randrot-method}
\title{Dimensions of an Object}
\usage{
\S4method{dim}{init.randrot}(x)

\S4method{dim}{init.batch.randrot}(x)
}
\arguments{
\item{x}{An object of class
\code{\link[randRotation:init.randrot-class]{init.randrot-class}} or
\code{\link[randRotation:init.batch.randrot-class]{init.batch.randrot-class}}.}
}
\value{
Vector of length two with number of \code{features} and number of
  \code{samples}. See also
  \code{\link[randRotation:init.randrot]{init.randrot}}.
}
\description{
Retrieve the dimensions of an object.
}
