% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randRot.R
\name{show,initRandrot-method}
\alias{show,initRandrot-method}
\alias{show,initBatchRandrot-method}
\alias{show,rotateStat-method}
\title{Show an Object}
\usage{
\S4method{show}{initRandrot}(object)

\S4method{show}{initBatchRandrot}(object)

\S4method{show}{rotateStat}(object)
}
\arguments{
\item{object}{An object of class
\code{\link[randRotation:initRandrot-class]{initRandrot-class}},
\code{\link[randRotation:initRandrot-class]{initRandrotW-class}} or
\code{\link[randRotation:initBatchRandrot-class]{initBatchRandrot-class}}.}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
Display the object by printing structured summary information.
}
\details{
The show method always displays the original design matrix
  (\code{X}), not the transformed (whitened) versions.
}
